+++ example/iw_database_example.dart
@@ -1,6 +1,0 @@
-import 'package:iw_database/iw_database.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}